[Refactor] Moved model interfaces in a dedicated package.
