+++ app/src/main/java/com/skichrome/oc/easyvgp/model/local/base/AdminSource.kt
@@ -1,11 +1,12 @@
-package com.skichrome.oc.easyvgp.model
+package com.skichrome.oc.easyvgp.model.local.base
 
 import androidx.lifecycle.LiveData
+import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.local.database.ControlPoint
 import com.skichrome.oc.easyvgp.model.local.database.MachineType
 import com.skichrome.oc.easyvgp.model.local.database.MachineTypeWithControlPoints
 
-interface AdminRepository
+interface AdminSource
 {
     fun observeMachineType(): LiveData<Results<List<MachineType>>>
     fun observeControlPoints(): LiveData<Results<List<ControlPoint>>>
