+++ app/src/main/java/com/skichrome/oc/easyvgp/model/local/base/CustomerRepository.kt
@@ -1,6 +1,7 @@
-package com.skichrome.oc.easyvgp.model
+package com.skichrome.oc.easyvgp.model.local.base
 
 import androidx.lifecycle.LiveData
+import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.local.database.Customer
 
 interface CustomerRepository
