+++ app/src/main/java/com/skichrome/oc/easyvgp/model/local/base/HomeRepository.kt
@@ -1,5 +1,6 @@
-package com.skichrome.oc.easyvgp.model
+package com.skichrome.oc.easyvgp.model.local.base
 
+import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.local.database.UserAndCompany
 
 interface HomeRepository
