+++ app/src/main/java/com/skichrome/oc/easyvgp/model/local/base/MachineRepository.kt
@@ -1,10 +1,11 @@
-package com.skichrome.oc.easyvgp.model
+package com.skichrome.oc.easyvgp.model.local.base
 
 import androidx.lifecycle.LiveData
+import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.local.database.Machine
 import com.skichrome.oc.easyvgp.model.local.database.MachineType
 
-interface MachineSource
+interface MachineRepository
 {
     fun observeMachines(): LiveData<Results<List<Machine>>>
     fun observeMachineTypes(): LiveData<Results<List<MachineType>>>
