+++ app/src/main/java/com/skichrome/oc/easyvgp/model/remote/CustomerRemoteRepository.kt
@@ -1,15 +1,16 @@
 package com.skichrome.oc.easyvgp.model.remote
 
 import androidx.lifecycle.LiveData
-import com.skichrome.oc.easyvgp.model.CustomerSource
 import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.Results.Error
+import com.skichrome.oc.easyvgp.model.local.base.CustomerSource
 import com.skichrome.oc.easyvgp.model.local.database.Customer
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
-class CustomerRemoteRepository(private val dispatcher: CoroutineDispatcher = Dispatchers.IO) : CustomerSource
+class CustomerRemoteRepository(private val dispatcher: CoroutineDispatcher = Dispatchers.IO) :
+    CustomerSource
 {
     override suspend fun saveCustomers(customers: Array<Customer>): Results<List<Long>> = withContext(dispatcher) {
         return@withContext Results.Success(listOf(customers.size.toLong()))
