+++ app/src/main/java/com/skichrome/oc/easyvgp/model/remote/RemoteHomeSource.kt
@@ -2,10 +2,10 @@ package com.skichrome.oc.easyvgp.model.remote
 
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ktx.toObject
-import com.skichrome.oc.easyvgp.model.HomeSource
 import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.Results.Error
 import com.skichrome.oc.easyvgp.model.Results.Success
+import com.skichrome.oc.easyvgp.model.local.base.HomeSource
 import com.skichrome.oc.easyvgp.model.local.database.ControlPoint
 import com.skichrome.oc.easyvgp.model.local.database.MachineType
 import com.skichrome.oc.easyvgp.model.local.database.MachineTypeWithControlPoints
@@ -16,7 +16,8 @@ import com.skichrome.oc.easyvgp.model.remote.util.RemoteMachineTypeWithControlPo
 import com.skichrome.oc.easyvgp.util.*
 import kotlinx.coroutines.*
 
-class RemoteHomeSource(private val dispatchers: CoroutineDispatcher = Dispatchers.IO) : HomeSource
+class RemoteHomeSource(private val dispatchers: CoroutineDispatcher = Dispatchers.IO) :
+    HomeSource
 {
     // =================================
     //              Fields
