+++ app/src/main/java/com/skichrome/oc/easyvgp/model/remote/RemoteVgpListSource.kt
@@ -13,9 +13,9 @@ import com.google.firebase.storage.ktx.storageMetadata
 import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.Results.Error
 import com.skichrome.oc.easyvgp.model.Results.Success
-import com.skichrome.oc.easyvgp.model.VgpListSource
 import com.skichrome.oc.easyvgp.model.local.ChoicePossibility
 import com.skichrome.oc.easyvgp.model.local.VerificationType
+import com.skichrome.oc.easyvgp.model.local.base.VgpListSource
 import com.skichrome.oc.easyvgp.model.local.database.*
 import com.skichrome.oc.easyvgp.model.remote.util.*
 import com.skichrome.oc.easyvgp.util.*
