+++ app/src/main/java/com/skichrome/oc/easyvgp/viewmodel/AdminViewModel.kt
@@ -4,9 +4,9 @@ import android.util.Log
 import androidx.databinding.ObservableBoolean
 import androidx.lifecycle.*
 import com.skichrome.oc.easyvgp.R
-import com.skichrome.oc.easyvgp.model.AdminRepository
 import com.skichrome.oc.easyvgp.model.Results.Error
 import com.skichrome.oc.easyvgp.model.Results.Success
+import com.skichrome.oc.easyvgp.model.local.base.AdminRepository
 import com.skichrome.oc.easyvgp.model.local.database.ControlPoint
 import com.skichrome.oc.easyvgp.model.local.database.MachineType
 import com.skichrome.oc.easyvgp.model.local.database.MachineTypeWithControlPoints
