+++ app/src/main/java/com/skichrome/oc/easyvgp/viewmodel/MachineViewModel.kt
@@ -2,8 +2,8 @@ package com.skichrome.oc.easyvgp.viewmodel
 
 import androidx.lifecycle.*
 import com.skichrome.oc.easyvgp.R
-import com.skichrome.oc.easyvgp.model.MachineRepository
 import com.skichrome.oc.easyvgp.model.Results.Success
+import com.skichrome.oc.easyvgp.model.local.base.MachineRepository
 import com.skichrome.oc.easyvgp.model.local.database.Machine
 import com.skichrome.oc.easyvgp.model.local.database.MachineType
 import com.skichrome.oc.easyvgp.util.Event
