+++ app/src/main/java/com/skichrome/oc/easyvgp/viewmodel/vmfactory/AdminViewModelFactory.kt
@@ -2,7 +2,7 @@ package com.skichrome.oc.easyvgp.viewmodel.vmfactory
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.skichrome.oc.easyvgp.model.AdminRepository
+import com.skichrome.oc.easyvgp.model.local.base.AdminRepository
 import com.skichrome.oc.easyvgp.viewmodel.AdminViewModel
 
 class AdminViewModelFactory(private val repository: AdminRepository) : ViewModelProvider.NewInstanceFactory()
