+++ app/src/main/java/com/skichrome/oc/easyvgp/viewmodel/vmfactory/CustomerViewModelFactory.kt
@@ -2,7 +2,7 @@ package com.skichrome.oc.easyvgp.viewmodel.vmfactory
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.skichrome.oc.easyvgp.model.CustomerRepository
+import com.skichrome.oc.easyvgp.model.local.base.CustomerRepository
 import com.skichrome.oc.easyvgp.viewmodel.CustomerViewModel
 
 class CustomerViewModelFactory(private val repository: CustomerRepository) : ViewModelProvider.NewInstanceFactory()
