+++ app/src/main/java/com/skichrome/oc/easyvgp/viewmodel/vmfactory/HomeViewModelFactory.kt
@@ -2,7 +2,7 @@ package com.skichrome.oc.easyvgp.viewmodel.vmfactory
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.skichrome.oc.easyvgp.model.HomeRepository
+import com.skichrome.oc.easyvgp.model.local.base.HomeRepository
 import com.skichrome.oc.easyvgp.viewmodel.HomeViewModel
 
 class HomeViewModelFactory(private val repository: HomeRepository) : ViewModelProvider.NewInstanceFactory()
