+++ app/src/main/java/com/skichrome/oc/easyvgp/viewmodel/vmfactory/VgpListViewModelFactory.kt
@@ -2,7 +2,7 @@ package com.skichrome.oc.easyvgp.viewmodel.vmfactory
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.skichrome.oc.easyvgp.model.VgpListRepository
+import com.skichrome.oc.easyvgp.model.local.base.VgpListRepository
 import com.skichrome.oc.easyvgp.viewmodel.VgpListViewModel
 
 class VgpListViewModelFactory(private val repository: VgpListRepository) : ViewModelProvider.NewInstanceFactory()
