+++ app/src/test/java/com/skichrome/oc/easyvgp/model/source/FakeCustomerSource.kt
@@ -2,10 +2,10 @@ package com.skichrome.oc.easyvgp.model.source
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.skichrome.oc.easyvgp.model.CustomerSource
 import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.Results.Error
 import com.skichrome.oc.easyvgp.model.Results.Success
+import com.skichrome.oc.easyvgp.model.local.base.CustomerSource
 import com.skichrome.oc.easyvgp.model.local.database.Customer
 
 class FakeCustomerSource(private var customerServiceData: LinkedHashMap<Long, Customer> = LinkedHashMap()) :
