+++ app/src/test/java/com/skichrome/oc/easyvgp/model/source/FakeMachineDataSource.kt
@@ -3,10 +3,10 @@ package com.skichrome.oc.easyvgp.model.source
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.map
-import com.skichrome.oc.easyvgp.model.MachineSource
 import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.Results.Error
 import com.skichrome.oc.easyvgp.model.Results.Success
+import com.skichrome.oc.easyvgp.model.local.base.MachineSource
 import com.skichrome.oc.easyvgp.model.local.database.Machine
 import com.skichrome.oc.easyvgp.model.local.database.MachineType
 import com.skichrome.oc.easyvgp.util.ItemNotFoundException
