+++ app/src/test/java/com/skichrome/oc/easyvgp/viewmodel/source/FakeCustomerRepository.kt
@@ -2,9 +2,9 @@ package com.skichrome.oc.easyvgp.viewmodel.source
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.skichrome.oc.easyvgp.model.CustomerRepository
 import com.skichrome.oc.easyvgp.model.Results
 import com.skichrome.oc.easyvgp.model.Results.Success
+import com.skichrome.oc.easyvgp.model.local.base.CustomerRepository
 import com.skichrome.oc.easyvgp.model.local.database.Customer
 
 class FakeCustomerRepository : CustomerRepository
